+++ tests/testthat/test-tbl-lazy.R
@@ -40,3 +40,7 @@ test_that("base source of tbl_lazy is always 'df'", {
   out <- lazy_frame(x = 1, y = 5) %>% sql_build()
   expect_equal(out, ident("df"))
 })
+
+test_that("names matches in-memory equivalent", {
+  expect_identical(names(lazy_frame(x = 1, y = 5)), c("x", "y"))
+})